requirements.txt
alembic==1.13.1
asyncpg==0.29.0
fastapi==0.109.2
pydantic==2.6.1
pydantic-settings==2.2.1
requests==2.31.0
SQLAlchemy==2.0.27
uvicorn==0.27.1
black==24.2.0
pylint==3.0.3
python-dotenv~=1.0.1
h11~=0.14.0
pip~=20.0.2
Mako~=1.3.2
MarkupSafe~=2.1.5
exceptiongroup~=1.2.0
sniffio~=1.3.0
platformdirs~=4.2.0
tomli~=2.0.1
packaging~=23.2
pathspec~=0.12.1
click~=8.1.7
certifi~=2024.2.2
starlette~=0.36.3
anyio~=4.3.0
urllib3~=2.2.1
greenlet~=3.0.3
idna~=3.6
setuptools~=44.0.0
zipp~=3.17.0